feat(vision): impl vision system & chore: restructure project
