--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/components/ServoPositionTuner.java
@@ -1,10 +1,0 @@
-package org.firstinspires.ftc.teamcode.components;
-
-import com.qualcomm.robotcore.hardware.ServoImplEx;
-
-public class ServoPositionTuner
-{
-    ServoImplEx clawGrab, clawRightRot, clawLeftRot, clawExtend;
-
-
-}
