+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/components/actuators/ServoPositionTuner.java
@@ -1,0 +1,10 @@
+package org.firstinspires.ftc.teamcode.components.actuators;
+
+import com.qualcomm.robotcore.hardware.ServoImplEx;
+
+public class ServoPositionTuner
+{
+    ServoImplEx clawGrab, clawRightRot, clawLeftRot, clawExtend;
+
+
+}
