+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/messages/commands/TankCommandMessage.java
@@ -1,0 +1,17 @@
+package org.firstinspires.ftc.teamcode.messages.commands;
+
+public final class TankCommandMessage
+{
+    public long timestamp;
+    public double voltage;
+    public double leftPower;
+    public double rightPower;
+
+    public TankCommandMessage(double voltage, double leftPower, double rightPower)
+    {
+        this.timestamp = System.nanoTime();
+        this.voltage = voltage;
+        this.leftPower = leftPower;
+        this.rightPower = rightPower;
+    }
+}
