+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/messages/inputs/PoseMessage.java
@@ -1,0 +1,20 @@
+package org.firstinspires.ftc.teamcode.messages.inputs;
+
+import com.acmerobotics.roadrunner.Pose2d;
+
+public final class PoseMessage
+{
+    public long timestamp;
+    public double x;
+    public double y;
+    public double heading;
+
+    public PoseMessage(Pose2d pose)
+    {
+        this.timestamp = System.nanoTime();
+        this.x = pose.position.x;
+        this.y = pose.position.y;
+        this.heading = pose.heading.toDouble();
+    }
+}
+
