+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/messages/inputs/TankLocalizerInputsMessage.java
@@ -1,0 +1,19 @@
+package org.firstinspires.ftc.teamcode.messages.inputs;
+
+import com.acmerobotics.roadrunner.ftc.PositionVelocityPair;
+
+import java.util.List;
+
+public final class TankLocalizerInputsMessage
+{
+    public long timestamp;
+    public PositionVelocityPair[] left;
+    public PositionVelocityPair[] right;
+
+    public TankLocalizerInputsMessage(List<PositionVelocityPair> left, List<PositionVelocityPair> right)
+    {
+        this.timestamp = System.nanoTime();
+        this.left = left.toArray(new PositionVelocityPair[0]);
+        this.right = right.toArray(new PositionVelocityPair[0]);
+    }
+}
