--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/parts/enums/ArmPosition.java
@@ -1,8 +1,0 @@
-package org.firstinspires.ftc.teamcode.parts.enums;
-
-public enum ArmPosition
-{
-    DETECTION,
-    GROUND,
-    BOARD
-}
