--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/parts/enums/ClawRotation.java
@@ -1,9 +1,0 @@
-package org.firstinspires.ftc.teamcode.parts.enums;
-
-public enum ClawRotation
-{
-    GROUND,
-    HIDDEN,
-    DETECTION,
-    BOARD,
-}
