--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/parts/enums/RobotState.java
@@ -1,10 +1,0 @@
-package org.firstinspires.ftc.teamcode.parts.enums;
-
-public enum RobotState
-{
-    IDLE,
-    DRIVING,
-    SLIDING,
-    ARM_CONTROL,
-    LAUNCHING, CLAW_CONTROL
-}
