+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tuning/LocalizationTest.java
@@ -8,9 +8,9 @@ import com.acmerobotics.roadrunner.PoseVelocity2d;
 import com.acmerobotics.roadrunner.Vector2d;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
-import org.firstinspires.ftc.teamcode.Drawing;
-import org.firstinspires.ftc.teamcode.MecanumDrive;
-import org.firstinspires.ftc.teamcode.TankDrive;
+import org.firstinspires.ftc.teamcode.utils.Drawing;
+import org.firstinspires.ftc.teamcode.drive.MecanumDrive;
+import org.firstinspires.ftc.teamcode.drive.TankDrive;
 
 public class LocalizationTest extends LinearOpMode
 {
@@ -27,13 +27,7 @@ public class LocalizationTest extends LinearOpMode
 
             while (opModeIsActive())
             {
-                drive.setDrivePowers(new PoseVelocity2d(
-                        new Vector2d(
-                                -gamepad1.left_stick_y,
-                                -gamepad1.left_stick_x
-                        ),
-                        -gamepad1.right_stick_x
-                ));
+                drive.setDrivePowers(new PoseVelocity2d(new Vector2d(-gamepad1.left_stick_y, -gamepad1.left_stick_x), -gamepad1.right_stick_x));
 
                 drive.updatePoseEstimate();
 
@@ -47,7 +41,8 @@ public class LocalizationTest extends LinearOpMode
                 Drawing.drawRobot(packet.fieldOverlay(), drive.pose);
                 FtcDashboard.getInstance().sendTelemetryPacket(packet);
             }
-        } else if (TuningOpModes.DRIVE_CLASS.equals(TankDrive.class))
+        }
+        else if (TuningOpModes.DRIVE_CLASS.equals(TankDrive.class))
         {
             TankDrive drive = new TankDrive(hardwareMap, new Pose2d(0, 0, 0));
 
@@ -55,13 +50,7 @@ public class LocalizationTest extends LinearOpMode
 
             while (opModeIsActive())
             {
-                drive.setDrivePowers(new PoseVelocity2d(
-                        new Vector2d(
-                                -gamepad1.left_stick_y,
-                                0.0
-                        ),
-                        -gamepad1.right_stick_x
-                ));
+                drive.setDrivePowers(new PoseVelocity2d(new Vector2d(-gamepad1.left_stick_y, 0.0), -gamepad1.right_stick_x));
 
                 drive.updatePoseEstimate();
 
@@ -75,7 +64,8 @@ public class LocalizationTest extends LinearOpMode
                 Drawing.drawRobot(packet.fieldOverlay(), drive.pose);
                 FtcDashboard.getInstance().sendTelemetryPacket(packet);
             }
-        } else
+        }
+        else
         {
             throw new RuntimeException();
         }
