+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tuning/ManualFeedbackTuner.java
@@ -4,10 +4,10 @@ import com.acmerobotics.roadrunner.Pose2d;
 import com.acmerobotics.roadrunner.ftc.Actions;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
-import org.firstinspires.ftc.teamcode.MecanumDrive;
-import org.firstinspires.ftc.teamcode.TankDrive;
-import org.firstinspires.ftc.teamcode.ThreeDeadWheelLocalizer;
-import org.firstinspires.ftc.teamcode.TwoDeadWheelLocalizer;
+import org.firstinspires.ftc.teamcode.drive.MecanumDrive;
+import org.firstinspires.ftc.teamcode.drive.TankDrive;
+import org.firstinspires.ftc.teamcode.drive.ThreeDeadWheelLocalizer;
+import org.firstinspires.ftc.teamcode.drive.TwoDeadWheelLocalizer;
 
 public final class ManualFeedbackTuner extends LinearOpMode
 {
@@ -22,13 +22,17 @@ public final class ManualFeedbackTuner extends LinearOpMode
 
             if (drive.localizer instanceof TwoDeadWheelLocalizer)
             {
-                if (TwoDeadWheelLocalizer.PARAMS.perpXTicks == 0 && TwoDeadWheelLocalizer.PARAMS.parYTicks == 0)
+                if (TwoDeadWheelLocalizer.PARAMS.perpXTicks == 0 &&
+                    TwoDeadWheelLocalizer.PARAMS.parYTicks == 0)
                 {
                     throw new RuntimeException("Odometry wheel locations not set! Run AngularRampLogger to tune them.");
                 }
-            } else if (drive.localizer instanceof ThreeDeadWheelLocalizer)
+            }
+            else if (drive.localizer instanceof ThreeDeadWheelLocalizer)
             {
-                if (ThreeDeadWheelLocalizer.PARAMS.perpXTicks == 0 && ThreeDeadWheelLocalizer.PARAMS.par0YTicks == 0 && ThreeDeadWheelLocalizer.PARAMS.par1YTicks == 1)
+                if (ThreeDeadWheelLocalizer.PARAMS.perpXTicks == 0 &&
+                    ThreeDeadWheelLocalizer.PARAMS.par0YTicks == 0 &&
+                    ThreeDeadWheelLocalizer.PARAMS.par1YTicks == 1)
                 {
                     throw new RuntimeException("Odometry wheel locations not set! Run AngularRampLogger to tune them.");
                 }
@@ -37,25 +41,26 @@ public final class ManualFeedbackTuner extends LinearOpMode
 
             while (opModeIsActive())
             {
-                Actions.runBlocking(
-                        drive.actionBuilder(new Pose2d(0, 0, 0))
-                                .lineToX(DISTANCE)
-                                .lineToX(0)
-                                .build());
+                Actions.runBlocking(drive.actionBuilder(new Pose2d(0, 0, 0)).lineToX(DISTANCE).lineToX(0).build());
             }
-        } else if (TuningOpModes.DRIVE_CLASS.equals(TankDrive.class))
+        }
+        else if (TuningOpModes.DRIVE_CLASS.equals(TankDrive.class))
         {
             TankDrive drive = new TankDrive(hardwareMap, new Pose2d(0, 0, 0));
 
             if (drive.localizer instanceof TwoDeadWheelLocalizer)
             {
-                if (TwoDeadWheelLocalizer.PARAMS.perpXTicks == 0 && TwoDeadWheelLocalizer.PARAMS.parYTicks == 0)
+                if (TwoDeadWheelLocalizer.PARAMS.perpXTicks == 0 &&
+                    TwoDeadWheelLocalizer.PARAMS.parYTicks == 0)
                 {
                     throw new RuntimeException("Odometry wheel locations not set! Run AngularRampLogger to tune them.");
                 }
-            } else if (drive.localizer instanceof ThreeDeadWheelLocalizer)
+            }
+            else if (drive.localizer instanceof ThreeDeadWheelLocalizer)
             {
-                if (ThreeDeadWheelLocalizer.PARAMS.perpXTicks == 0 && ThreeDeadWheelLocalizer.PARAMS.par0YTicks == 0 && ThreeDeadWheelLocalizer.PARAMS.par1YTicks == 1)
+                if (ThreeDeadWheelLocalizer.PARAMS.perpXTicks == 0 &&
+                    ThreeDeadWheelLocalizer.PARAMS.par0YTicks == 0 &&
+                    ThreeDeadWheelLocalizer.PARAMS.par1YTicks == 1)
                 {
                     throw new RuntimeException("Odometry wheel locations not set! Run AngularRampLogger to tune them.");
                 }
@@ -64,13 +69,10 @@ public final class ManualFeedbackTuner extends LinearOpMode
 
             while (opModeIsActive())
             {
-                Actions.runBlocking(
-                        drive.actionBuilder(new Pose2d(0, 0, 0))
-                                .lineToX(DISTANCE)
-                                .lineToX(0)
-                                .build());
+                Actions.runBlocking(drive.actionBuilder(new Pose2d(0, 0, 0)).lineToX(DISTANCE).lineToX(0).build());
             }
-        } else
+        }
+        else
         {
             throw new RuntimeException();
         }
