+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tuning/SplineTest.java
@@ -5,8 +5,8 @@ import com.acmerobotics.roadrunner.Vector2d;
 import com.acmerobotics.roadrunner.ftc.Actions;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
-import org.firstinspires.ftc.teamcode.MecanumDrive;
-import org.firstinspires.ftc.teamcode.TankDrive;
+import org.firstinspires.ftc.teamcode.drive.MecanumDrive;
+import org.firstinspires.ftc.teamcode.drive.TankDrive;
 
 public final class SplineTest extends LinearOpMode
 {
@@ -20,23 +20,19 @@ public final class SplineTest extends LinearOpMode
 
             waitForStart();
 
-            Actions.runBlocking(
-                    drive.actionBuilder(beginPose)
-                            .splineTo(new Vector2d(30, 30), Math.PI / 2)
-                            .splineTo(new Vector2d(0, 60), Math.PI)
-                            .build());
-        } else if (TuningOpModes.DRIVE_CLASS.equals(TankDrive.class))
+            Actions.runBlocking(drive.actionBuilder(beginPose).splineTo(new Vector2d(30, 30),
+                    Math.PI / 2).splineTo(new Vector2d(0, 60), Math.PI).build());
+        }
+        else if (TuningOpModes.DRIVE_CLASS.equals(TankDrive.class))
         {
             TankDrive drive = new TankDrive(hardwareMap, beginPose);
 
             waitForStart();
 
-            Actions.runBlocking(
-                    drive.actionBuilder(beginPose)
-                            .splineTo(new Vector2d(30, 30), Math.PI / 2)
-                            .splineTo(new Vector2d(0, 60), Math.PI)
-                            .build());
-        } else
+            Actions.runBlocking(drive.actionBuilder(beginPose).splineTo(new Vector2d(30, 30),
+                    Math.PI / 2).splineTo(new Vector2d(0, 60), Math.PI).build());
+        }
+        else
         {
             throw new RuntimeException();
         }
