+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tuning/TuningOpModes.java
@@ -22,10 +22,10 @@ import com.qualcomm.robotcore.eventloop.opmode.OpModeManager;
 import com.qualcomm.robotcore.eventloop.opmode.OpModeRegistrar;
 
 import org.firstinspires.ftc.robotcore.internal.opmode.OpModeMeta;
-import org.firstinspires.ftc.teamcode.MecanumDrive;
-import org.firstinspires.ftc.teamcode.TankDrive;
-import org.firstinspires.ftc.teamcode.ThreeDeadWheelLocalizer;
-import org.firstinspires.ftc.teamcode.TwoDeadWheelLocalizer;
+import org.firstinspires.ftc.teamcode.drive.MecanumDrive;
+import org.firstinspires.ftc.teamcode.drive.TankDrive;
+import org.firstinspires.ftc.teamcode.drive.ThreeDeadWheelLocalizer;
+import org.firstinspires.ftc.teamcode.drive.TwoDeadWheelLocalizer;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -45,17 +45,14 @@ public final class TuningOpModes
 
     private static OpModeMeta metaForClass(Class<? extends OpMode> cls)
     {
-        return new OpModeMeta.Builder()
-                .setName(cls.getSimpleName())
-                .setGroup(GROUP)
-                .setFlavor(OpModeMeta.Flavor.TELEOP)
-                .build();
+        return new OpModeMeta.Builder().setName(cls.getSimpleName()).setGroup(GROUP).setFlavor(OpModeMeta.Flavor.TELEOP).build();
     }
 
     @OpModeRegistrar
     public static void register(OpModeManager manager)
     {
-        if (DISABLED) return;
+        if (DISABLED)
+        { return; }
 
         DriveViewFactory dvf;
         if (DRIVE_CLASS.equals(MecanumDrive.class))
@@ -73,49 +70,32 @@ public final class TuningOpModes
                     leftEncs.add(dl.leftBack);
                     rightEncs.add(dl.rightFront);
                     rightEncs.add(dl.rightBack);
-                } else if (md.localizer instanceof ThreeDeadWheelLocalizer)
+                }
+                else if (md.localizer instanceof ThreeDeadWheelLocalizer)
                 {
                     ThreeDeadWheelLocalizer dl = (ThreeDeadWheelLocalizer) md.localizer;
                     parEncs.add(dl.par0);
                     parEncs.add(dl.par1);
                     perpEncs.add(dl.perp);
-                } else if (md.localizer instanceof TwoDeadWheelLocalizer)
+                }
+                else if (md.localizer instanceof TwoDeadWheelLocalizer)
                 {
                     TwoDeadWheelLocalizer dl = (TwoDeadWheelLocalizer) md.localizer;
                     parEncs.add(dl.par);
                     perpEncs.add(dl.perp);
-                } else
+                }
+                else
                 {
-                    throw new RuntimeException("unknown localizer: " + md.localizer.getClass().getName());
+                    throw new RuntimeException(
+                            "unknown localizer: " + md.localizer.getClass().getName());
                 }
 
-                return new DriveView(
-                        DriveType.MECANUM,
-                        MecanumDrive.PARAMS.inPerTick,
-                        MecanumDrive.PARAMS.maxWheelVel,
-                        MecanumDrive.PARAMS.minProfileAccel,
-                        MecanumDrive.PARAMS.maxProfileAccel,
-                        hardwareMap.getAll(LynxModule.class),
-                        Arrays.asList(
-                                md.leftFront,
-                                md.leftBack
-                        ),
-                        Arrays.asList(
-                                md.rightFront,
-                                md.rightBack
-                        ),
-                        leftEncs,
-                        rightEncs,
-                        parEncs,
-                        perpEncs,
-                        md.lazyImu,
-                        md.voltageSensor,
-                        () -> new MotorFeedforward(MecanumDrive.PARAMS.kS,
-                                MecanumDrive.PARAMS.kV / MecanumDrive.PARAMS.inPerTick,
-                                MecanumDrive.PARAMS.kA / MecanumDrive.PARAMS.inPerTick)
-                );
+                return new DriveView(DriveType.MECANUM, MecanumDrive.PARAMS.inPerTick, MecanumDrive.PARAMS.maxWheelVel, MecanumDrive.PARAMS.minProfileAccel, MecanumDrive.PARAMS.maxProfileAccel, hardwareMap.getAll(LynxModule.class), Arrays.asList(md.leftFront, md.leftBack), Arrays.asList(md.rightFront, md.rightBack), leftEncs, rightEncs, parEncs, perpEncs, md.lazyImu, md.voltageSensor, () -> new MotorFeedforward(MecanumDrive.PARAMS.kS,
+                        MecanumDrive.PARAMS.kV / MecanumDrive.PARAMS.inPerTick,
+                        MecanumDrive.PARAMS.kA / MecanumDrive.PARAMS.inPerTick));
             };
-        } else if (DRIVE_CLASS.equals(TankDrive.class))
+        }
+        else if (DRIVE_CLASS.equals(TankDrive.class))
         {
             dvf = hardwareMap ->
             {
@@ -128,43 +108,32 @@ public final class TuningOpModes
                     TankDrive.DriveLocalizer dl = (TankDrive.DriveLocalizer) td.localizer;
                     leftEncs.addAll(dl.leftEncs);
                     rightEncs.addAll(dl.rightEncs);
-                } else if (td.localizer instanceof ThreeDeadWheelLocalizer)
+                }
+                else if (td.localizer instanceof ThreeDeadWheelLocalizer)
                 {
                     ThreeDeadWheelLocalizer dl = (ThreeDeadWheelLocalizer) td.localizer;
                     parEncs.add(dl.par0);
                     parEncs.add(dl.par1);
                     perpEncs.add(dl.perp);
-                } else if (td.localizer instanceof TwoDeadWheelLocalizer)
+                }
+                else if (td.localizer instanceof TwoDeadWheelLocalizer)
                 {
                     TwoDeadWheelLocalizer dl = (TwoDeadWheelLocalizer) td.localizer;
                     parEncs.add(dl.par);
                     perpEncs.add(dl.perp);
-                } else
+                }
+                else
                 {
-                    throw new RuntimeException("unknown localizer: " + td.localizer.getClass().getName());
+                    throw new RuntimeException(
+                            "unknown localizer: " + td.localizer.getClass().getName());
                 }
 
-                return new DriveView(
-                        DriveType.TANK,
-                        TankDrive.PARAMS.inPerTick,
-                        TankDrive.PARAMS.maxWheelVel,
-                        TankDrive.PARAMS.minProfileAccel,
-                        TankDrive.PARAMS.maxProfileAccel,
-                        hardwareMap.getAll(LynxModule.class),
-                        td.leftMotors,
-                        td.rightMotors,
-                        leftEncs,
-                        rightEncs,
-                        parEncs,
-                        perpEncs,
-                        td.lazyImu,
-                        td.voltageSensor,
-                        () -> new MotorFeedforward(TankDrive.PARAMS.kS,
-                                TankDrive.PARAMS.kV / TankDrive.PARAMS.inPerTick,
-                                TankDrive.PARAMS.kA / TankDrive.PARAMS.inPerTick)
-                );
+                return new DriveView(DriveType.TANK, TankDrive.PARAMS.inPerTick, TankDrive.PARAMS.maxWheelVel, TankDrive.PARAMS.minProfileAccel, TankDrive.PARAMS.maxProfileAccel, hardwareMap.getAll(LynxModule.class), td.leftMotors, td.rightMotors, leftEncs, rightEncs, parEncs, perpEncs, td.lazyImu, td.voltageSensor, () -> new MotorFeedforward(TankDrive.PARAMS.kS,
+                        TankDrive.PARAMS.kV / TankDrive.PARAMS.inPerTick,
+                        TankDrive.PARAMS.kA / TankDrive.PARAMS.inPerTick));
             };
-        } else
+        }
+        else
         {
             throw new RuntimeException();
         }
@@ -184,14 +153,7 @@ public final class TuningOpModes
 
         FtcDashboard.getInstance().withConfigRoot(configRoot ->
         {
-            for (Class<?> c : Arrays.asList(
-                    AngularRampLogger.class,
-                    ForwardRampLogger.class,
-                    LateralRampLogger.class,
-                    ManualFeedforwardTuner.class,
-                    MecanumMotorDirectionDebugger.class,
-                    ManualFeedbackTuner.class
-            ))
+            for (Class<?> c : Arrays.asList(AngularRampLogger.class, ForwardRampLogger.class, LateralRampLogger.class, ManualFeedforwardTuner.class, MecanumMotorDirectionDebugger.class, ManualFeedbackTuner.class))
             {
                 configRoot.putVariable(c.getSimpleName(), ReflectionConfig.createVariableFromClass(c));
             }
