+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/types/ArmPosition.java
@@ -1,0 +1,6 @@
+package org.firstinspires.ftc.teamcode.types;
+
+public enum ArmPosition
+{
+    DETECTION, GROUND, BOARD
+}
