+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/types/ClawRotation.java
@@ -1,0 +1,6 @@
+package org.firstinspires.ftc.teamcode.types;
+
+public enum ClawRotation
+{
+    GROUND, HIDDEN, DETECTION, BOARD,
+}
