+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/types/RobotState.java
@@ -1,0 +1,6 @@
+package org.firstinspires.ftc.teamcode.types;
+
+public enum RobotState
+{
+    IDLE, DRIVING, SLIDING, ARM_CONTROL, LAUNCHING, CLAW_CONTROL
+}
