+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/utils/Localizer.java
@@ -1,0 +1,9 @@
+package org.firstinspires.ftc.teamcode.utils;
+
+import com.acmerobotics.roadrunner.Time;
+import com.acmerobotics.roadrunner.Twist2dDual;
+
+public interface Localizer
+{
+    Twist2dDual<Time> update();
+}
